Revert "NVMe bugfix: Use integer math and return type for nvmeGetIOSlip() so that {0, 0} initial state returns -1, which is a safe value if disk_read() occurs first."
This reverts commit fccc6f65d81dd72d186ace58d583c617dd764ad1.

+++ base_cmd.vitis/WAVE/src/nvme.c
@@ -335,9 +335,9 @@ int nvmeServiceIOCompletions(u16 maxCompletions)
 	return numCompletions;
 }
 
-int nvmeGetIOSlip(void)
+u16 nvmeGetIOSlip(void)
 {
-	return ((int)io_cid - (int)io_cid_last_completed - 1);
+	return (u16)(io_cid - io_cid_last_completed - 1);
 }
 
 // Private Function Definitions ----------------------------------------------------------------------------------------
